// src/modules/auth/auth.enum.ts
export enum AuthErrorCode {
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  EMAIL_NOT_VERIFIED = 'EMAIL_NOT_VERIFIED',
  EMAIL_ALREADY_IN_USE = 'EMAIL_ALREADY_IN_USE',
  REFRESH_TOKEN_INVALID = 'REFRESH_TOKEN_INVALID',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  VERIFICATION_LINK_INVALID = 'VERIFICATION_LINK_INVALID',
  RESET_LINK_INVALID = 'RESET_LINK_INVALID',
  NO_PASSWORD_SET = 'NO_PASSWORD_SET',
  SAME_PASSWORD = 'SAME_PASSWORD',
  FORBIDDEN_ROLE = 'FORBIDDEN_ROLE',
  ACCOUNT_BANNED = 'ACCOUNT_BANNED',
}